[quant][pt2e] Add `model_is_exported` util function
Summary: This commit adds the `model_is_exported` util function
for users to be able to easily tell what APIs to call to move
their models between train and eval modes. This has the
additional advantage of hiding the implementation of how we
detect a model is exported, in case the metadata format changes
in the future.

Test Plan:
python test/test_quantization.py TestQuantizePT2E.test_model_is_exported

ghstack-source-id: 1fd1c11e6261207cd01df03ab0ce471fd23d5a1a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119726